refactor(app): split chat state transitions
