+++ apps/app/src/features/chat/runtime/chat-outgoing.ts
@@ -1,0 +1,56 @@
+import type { PromptPart } from "@vibest/contract";
+import type { UIMessage } from "ai";
+
+import type { ChatDraft, ChatEffects } from "./chat-runtime-types";
+import { addUnique, isChatActive, type ChatState, type OutgoingMessage } from "./chat-state";
+
+export const sendingMessage = (state: ChatState): OutgoingMessage | undefined =>
+  state.outgoing.find((message) => message.status === "sending");
+
+const toUserMessage = (messageId: string, parts: ReadonlyArray<PromptPart>): UIMessage => ({
+  id: messageId,
+  role: "user",
+  parts: parts.map((part) =>
+    part.type === "data-inspector" ? { type: "data-inspector", data: part.data } : part,
+  ) as UIMessage["parts"],
+});
+
+export function pushUserMessage(
+  state: ChatDraft,
+  messageId: string,
+  parts: ReadonlyArray<PromptPart>,
+): boolean {
+  if (state.session.messages.some((message) => message.id === messageId)) return false;
+  state.session.messages.push(toUserMessage(messageId, parts));
+  state.session.error = undefined;
+  return true;
+}
+
+export function removeUserMessage(state: ChatDraft, messageId: string): void {
+  state.session.messages = state.session.messages.filter((message) => message.id !== messageId);
+}
+
+export function maybeDispatchPrompt(state: ChatDraft, effects: ChatEffects): void {
+  const nextIndex = state.outgoing.findIndex((message) => message.status === "queued");
+  if (
+    !isChatActive(state) ||
+    sendingMessage(state) !== undefined ||
+    !state.sync.historyLoaded ||
+    state.sync.floor !== null ||
+    !state.prompt.boundaryOpen ||
+    nextIndex === -1
+  ) {
+    return;
+  }
+  const next = state.outgoing[nextIndex]!;
+  state.outgoing[nextIndex] = { ...next, status: "sending" };
+  if (!state.session.messages.some((message) => message.id === next.message.id)) {
+    state.session.messages.push(next.message);
+  }
+  state.session.status = "submitted";
+  state.session.error = undefined;
+  addUnique(state.prompt.pendingMessageIds, next.message.id);
+  state.prompt.boundaryOpen = false;
+  state.prompt.revision += 1;
+  effects.push({ type: "submitPrompt", messageId: next.message.id, parts: next.parts });
+}
